requirements.txt
transformers==4.3.2
gitpython==3.1.12
pytorch_lightning==1.1.3
rouge-score==0.0.4
sacrebleu==1.4.3
tensorflow-tensorboard==1.5.1
tensorboardX==2.1
tensorboard==2.4.1
sentencepiece==0.1.95
datasets==1.3.0
spacy
matplotlib